Most files: Move files into subdirectories
Almost all the files have been moved into folders and subfolders to keep everything organized and clean.
