+++ src/headers/switcher-data-structs.hpp
