+++ src/linux/advanced-scene-switcher-nix.cpp
