+++ src/osx/advanced-scene-switcher-osx.mm
