+++ src/priority.cpp
@@ -1,5 +1,5 @@
 #include <algorithm>
-#include "advanced-scene-switcher.hpp"
+#include "headers/advanced-scene-switcher.hpp"
 
 void SceneSwitcher::on_priorityUp_clicked()
 {
