+++ src/scene-transitions.cpp
@@ -1,4 +1,4 @@
-#include "advanced-scene-switcher.hpp"
+#include "headers/advanced-scene-switcher.hpp"
 
 void SceneSwitcher::on_transitionsAdd_clicked()
 {
