+++ src/win/advanced-scene-switcher-win.cpp
